requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.2.0